timer b testing up against the datasheet
The timers are getting closer to the description in the datasheet. I
have learned that using variables in processes can help detect a change
in a signal without causing delay, and without having to have a
old_signal as signal.  The todo file list has a lot of copy paste from
the datasheet regarding the timer, and stuff to test.