Improving the Timers
The code from 2006 has internal tri-state bus. That's bad for FPGA.  I'm
giving back the modules separate in and out ports. The Timer testbench
did not use the bus_proc procedures from the main test-bench. I've put
the procedures in a base_pck to use in all the test benches.
And then there are the anoying use of logics on the async resets.
Todo: Make all registers syncronous to the clock, and not some async
regs clocked to TIMERCLOCK. Also I have logic on the clock inputs. Its a
no-no !!!!!!